+++ visidata/loaders/archive.py
@@ -2,6 +2,7 @@ import pathlib
 import tarfile
 import zipfile
 import datetime
+import os.path
 from visidata.loaders import unzip_http
 
 from visidata import vd, VisiData, asyncthread, Sheet, Progress, Menu, options
@@ -84,13 +85,13 @@ Commands:
             return vd.openSource(Path(fi.filename, fp=fp, filesize=fi.file_size), filetype=options.filetype)
 
     def extract(self, *rows, path=None):
-        path = path or pathlib.Path('.')
+        path = path or Path('.')
 
         files = []
         for row in rows:
             r, _ = row
             vd.confirmOverwrite(path/r.filename)  #1452
-            self.extract_async(row)
+            self.extract_async(row, path=path)
 
     def sysopen_row(self, row):
         'Extract file in row to tempdir and launch $EDITOR.  Modifications will be discarded.'
@@ -112,6 +113,12 @@ Commands:
             if '://' in str(self.source):
                 unzip_http.warning = vd.warning
                 self._zfp = unzip_http.RemoteZipFile(str(self.source))
+            elif isinstance(self.source, Path):
+                if self.source.has_fp():  #when opening a zip inside tar or zip
+                    fp = self.source.open('rb')
+                else:
+                    fp = self.source
+                self._zfp = zipfile.ZipFile(fp, 'r')
             else:
                 self._zfp = zipfile.ZipFile(str(self.source), 'r')
 
@@ -122,18 +129,35 @@ Commands:
             yield [zi, Path(zi.filename)]
 
 
+#from https://docs.python.org/3/library/tarfile.html#tarfile.REGTYPE
+tarfile_type_names = {
+    tarfile.REGTYPE:"file",
+    tarfile.AREGTYPE:"file",
+    tarfile.LNKTYPE:"hard link",
+    tarfile.SYMTYPE:"symbolic link",
+    tarfile.CHRTYPE:"character device",
+    tarfile.BLKTYPE:"block device",
+    tarfile.DIRTYPE:"directory",
+    tarfile.FIFOTYPE:"FIFO",
+    tarfile.CONTTYPE:"contiguous file",
+    tarfile.GNUTYPE_LONGNAME:"GNU tar longname",
+    tarfile.GNUTYPE_LONGLINK:"GNU tar longlink",
+    tarfile.GNUTYPE_SPARSE:"GNU tar sparse file",
+}
 class TarSheet(Sheet):
     'Wrapper for `tarfile` library.'
     rowtype = 'files' # rowdef TarInfo
     columns = [
         ColumnAttr('name'),
+        Column('ext', getter=lambda col,row: row.isdir() and '/' or os.path.splitext(row.name)[1][1:]),
         ColumnAttr('size', type=int),
         ColumnAttr('mtime', type=date),
-        ColumnAttr('type', type=int),
+        Column('type', getter=lambda col, row: tarfile_type_names.get(row.type, 'unknown')),
         ColumnAttr('mode', type=int),
         ColumnAttr('uname'),
         ColumnAttr('gname')
     ]
+    nKeys=1
 
     def openRow(self, fi):
             tfp = tarfile.open(name=str(self.source))