+++ visidata/loaders/csv.py
@@ -13,6 +13,13 @@ vd.option('safety_first', False, 'sanitize input/output to handle edge cases, wi
 
 
 @VisiData.api
+def guess_csv_delimiter(vd, p):
+    'If csv_delimiter option has been modified from default, assume CSV format.'
+    
+    if vd.options.csv_delimiter != vd.options.getdefault('csv_delimiter'):
+        return dict(filetype='csv', _likelihood=2)
+
+@VisiData.api
 def guess_csv(vd, p):
     import csv
     csv.field_size_limit(2**31-1)  #288 Windows has max 32-bit