+++ visidata/mouse.py
@@ -13,6 +13,8 @@ BaseSheet.init('mouseY', int)
 
 @VisiData.after
 def initCurses(vd):
+    if not getattr(curses, 'mousemask', None):
+      return
     curses.MOUSE_ALL = 0xffffffff
     curses.mousemask(curses.MOUSE_ALL if vd.options.mouse_interval else 0)
     curses.def_prog_mode()
@@ -136,5 +138,5 @@ def go_mouse(sheet):
 
 Sheet.addCommand(None, 'scroll-mouse', 'sheet.topRowIndex=cursorRowIndex-mouseY+1', 'scroll to mouse cursor location')
 
-Sheet.addCommand('ScrollUp', 'scroll-up', 'cursorDown(options.scroll_incr); sheet.topRowIndex += options.scroll_incr', 'scroll one row up')
-Sheet.addCommand('ScrollDown', 'scroll-down', 'cursorDown(-options.scroll_incr); sheet.topRowIndex -= options.scroll_incr', 'scroll one row down')
+Sheet.addCommand('ScrollUp', 'scroll-up', 'cursorDown(options.scroll_incr); sheet.topRowIndex += options.scroll_incr', 'scroll up by one row increment')
+Sheet.addCommand('ScrollDown', 'scroll-down', 'cursorDown(-options.scroll_incr); sheet.topRowIndex -= options.scroll_incr', 'scroll down by one row increment')