+++ visidata/settings.py
@@ -338,6 +338,18 @@ def addCommand(cls, keystrokes, longname, execstr, helpstr='', replay=True, **kw
         vd.bindkey(keystrokes, longname, cls)
     return longname
 
+@BaseSheet.class_api
+@classmethod
+def removeCommand(cls, keystrokes, longname):
+    '''Remove a command from *cls* sheet type.
+
+    - *keystrokes*: if provided, unbind this specific keystroke.
+    - *longname*: name of the command to remove.
+    '''
+    vd.commands.unset(longname, cls)
+    if keystrokes:
+        vd.unbindkey(keystrokes, cls)
+
 def _command(cls, binding, longname, helpstr, **kwargs):
     def decorator(func):
         funcname = longname.replace('-', '_')
@@ -383,8 +395,12 @@ def getCommand(sheet, cmd):
         return cmd
 
     longname = cmd
+    seen = []
     while vd.bindkeys._get(longname, obj=sheet) is not None:
         longname = vd.bindkeys._get(longname, obj=sheet)
+        if longname in seen:
+            vd.fail(f'keystroke/command definitions form a cycle: {longname}')
+        seen.append(longname)
 
     return vd.commands._get(longname, obj=sheet)
 
@@ -447,12 +463,16 @@ def loadConfigAndPlugins(vd, args=AttrDict()):
     args_plugins_autoload = args.plugins_autoload if 'plugins_autoload' in args else True
     if not args.nothing and args_plugins_autoload and vd.options.plugins_autoload:
         from importlib.metadata import entry_points
+        eps_visidata = []
         try:
             eps = entry_points()
-            eps_visidata = eps.select(group='visidata.plugins') if 'visidata.plugins' in eps.groups else []
+            vp = 'visidata.plugins'
+            if hasattr(eps, 'groups'): #Python >= 3.10
+                eps_visidata = eps.select(group=vp)
+            else:                      #Python <  3.10
+                eps_visidata = eps.get(vp, [])
         except Exception as e:
-            eps_visidata = []
-            vd.warning('plugin autoload failed; see issue #1529')
+            vd.warning(f'plugin autoload failed; see issue #1529:  {e}')
 
         for ep in eps_visidata:
             try: