+++ visidata/undo.py
@@ -42,7 +42,8 @@ def undo(vd, sheet):
 
     cmdlogrows = itertools.dropwhile(lambda r: r.longname == 'set-option', sheet.cmdlog_sheet.rows)
     # skip the first remaining command, to exclude it from undo,
-    # because it is always the one that created the sheet
+    # because it is always the command that created the sheet, or for
+    # replayed cmdlogs, a no-op placeholder row
     for i, cmdlogrow in enumerate(reversed(list(cmdlogrows)[1:])):
         if cmdlogrow.undofuncs:
             for undofunc, args, kwargs, in cmdlogrow.undofuncs[::-1]:
@@ -51,7 +52,7 @@ def undo(vd, sheet):
             row_idx = len(sheet.cmdlog_sheet.rows)-1 - i
             del sheet.cmdlog_sheet.rows[row_idx]
 
-            sheet.calcColLayout()  # undofunc can invalidate the column layout and/or drawcache
+            vd.clearCaches()  # undofunc can invalidate the drawcache
 
             vd.moveToReplayContext(cmdlogrow, sheet)
             vd.status("%s undone" % cmdlogrow.longname)